+++ wren-ai-service/src/pipelines/common.py
@@ -1,3 +1,4 @@
+import asyncio
 import logging
 from pprint import pformat
 from typing import Any, Dict, List, Optional
@@ -16,7 +17,7 @@ logger = logging.getLogger("wren-ai-service")
 
 
 @component
-class GenerationPostProcessor:
+class SQLBreakdownGenerationPostProcessor:
     def __init__(self, engine: Engine):
         self._engine = engine
 
@@ -53,8 +54,8 @@ class GenerationPostProcessor:
                 }
 
         sql = self._build_cte_query(steps)
-        logger.debug(f"GenerationPostProcessor: steps: {pformat(steps)}")
-        logger.debug(f"GenerationPostProcessor: final sql: {sql}")
+        logger.debug(f"SQLBreakdownGenerationPostProcessor: steps: {pformat(steps)}")
+        logger.debug(f"SQLBreakdownGenerationPostProcessor: final sql: {sql}")
 
         if not await self._check_if_sql_executable(sql, project_id=project_id):
             return {
@@ -96,3 +97,90 @@ class GenerationPostProcessor:
             logger.exception(f"SQL is not executable: {error}")
 
         return status
+
+
+@component
+class SQLGenerationPostProcessor:
+    def __init__(self, engine: Engine):
+        self._engine = engine
+
+    @component.output_types(
+        valid_generation_results=List[Optional[Dict[str, Any]]],
+        invalid_generation_results=List[Optional[Dict[str, Any]]],
+    )
+    async def run(
+        self,
+        replies: List[str],
+        project_id: str | None = None,
+    ) -> dict:
+        try:
+            cleaned_generation_result = orjson.loads(
+                clean_generation_result(replies[0])
+            )["results"]
+
+            if isinstance(cleaned_generation_result, dict):
+                cleaned_generation_result = [cleaned_generation_result]
+
+            (
+                valid_generation_results,
+                invalid_generation_results,
+            ) = await self._classify_invalid_generation_results(
+                cleaned_generation_result, project_id=project_id
+            )
+
+            return {
+                "valid_generation_results": valid_generation_results,
+                "invalid_generation_results": invalid_generation_results,
+            }
+        except Exception as e:
+            logger.exception(f"Error in SQLGenerationPostProcessor: {e}")
+
+            return {
+                "valid_generation_results": [],
+                "invalid_generation_results": [],
+            }
+
+    async def _classify_invalid_generation_results(
+        self, generation_results: List[Dict[str, str]], project_id: str | None = None
+    ) -> List[Optional[Dict[str, str]]]:
+        valid_generation_results = []
+        invalid_generation_results = []
+
+        async def _task(result: Dict[str, str]):
+            quoted_sql, no_error = add_quotes(result["sql"])
+
+            if no_error:
+                status, _, error = await self._engine.execute_sql(
+                    quoted_sql, session, project_id=project_id
+                )
+
+                if status:
+                    valid_generation_results.append(
+                        {
+                            "sql": quoted_sql,
+                        }
+                    )
+                else:
+                    invalid_generation_results.append(
+                        {
+                            "sql": quoted_sql,
+                            "type": "DRY_RUN",
+                            "error": error,
+                        }
+                    )
+            else:
+                invalid_generation_results.append(
+                    {
+                        "sql": result["sql"],
+                        "type": "ADD_QUOTES",
+                        "error": "add_quotes failed",
+                    }
+                )
+
+        async with aiohttp.ClientSession() as session:
+            tasks = [
+                _task(generation_result) for generation_result in generation_results
+            ]
+            await asyncio.gather(*tasks)
+
+        return valid_generation_results, invalid_generation_results
