+++ wren-ai-service/src/pipelines/sql_expansion/generation.py
@@ -12,9 +12,7 @@ from langfuse.decorators import observe
 from src.core.engine import Engine
 from src.core.pipeline import BasicPipeline
 from src.core.provider import LLMProvider
-from src.pipelines.sql_generation.components.post_processors import (
-    GenerationPostProcessor,
-)
+from src.pipelines.common import SQLBreakdownGenerationPostProcessor
 from src.utils import async_timer, timer
 from src.web.v1.services.ask import AskHistory
 
@@ -54,7 +52,7 @@ async def generate_sql_expansion(prompt: dict, generator: Any) -> dict:
 @observe(capture_input=False)
 async def post_process(
     generate_sql_expansion: dict,
-    post_processor: GenerationPostProcessor,
+    post_processor: SQLBreakdownGenerationPostProcessor,
     project_id: str | None = None,
 ) -> dict:
     logger.debug(
@@ -81,7 +79,7 @@ class Generation(BasicPipeline):
             "prompt_builder": PromptBuilder(
                 template=sql_expansion_user_prompt_template
             ),
-            "post_processor": GenerationPostProcessor(engine=engine),
+            "post_processor": SQLBreakdownGenerationPostProcessor(engine=engine),
         }
 
         super().__init__(
