+++ wren-ai-service/src/pipelines/sql_generation/generation.py
@@ -12,9 +12,7 @@ from langfuse.decorators import observe
 from src.core.engine import Engine
 from src.core.pipeline import BasicPipeline
 from src.core.provider import LLMProvider
-from src.pipelines.sql_generation.components.post_processors import (
-    GenerationPostProcessor,
-)
+from src.pipelines.common import SQLGenerationPostProcessor
 from src.pipelines.sql_generation.components.prompts import (
     TEXT_TO_SQL_RULES,
     text_to_sql_system_prompt,
@@ -121,7 +119,7 @@ async def generate_sql(prompt: dict, generator: Any) -> dict:
 @observe(capture_input=False)
 async def post_process(
     generate_sql: dict,
-    post_processor: GenerationPostProcessor,
+    post_processor: SQLGenerationPostProcessor,
     project_id: str | None = None,
 ) -> dict:
     logger.debug(
@@ -144,7 +142,7 @@ class Generation(BasicPipeline):
                 system_prompt=text_to_sql_system_prompt
             ),
             "prompt_builder": PromptBuilder(template=text_to_sql_user_prompt_template),
-            "post_processor": GenerationPostProcessor(engine=engine),
+            "post_processor": SQLGenerationPostProcessor(engine=engine),
         }
 
         self._configs = {
